Use 'artist' not 'author' for the dc:author tag.
This code came from speexenc, which has always used
'author' for authorship metadata. This is the name
used by the Dublic Core tag set, but Vorbis and FLAC
have been using 'artist' instead for over a decade.

'Artist' isn't ideal outside of the original domain
of popular music, but sticking with it is less
surprising than making a different choice now.