Try linking opus when testing -fPIE.
When build against a static libopus (built without -fPIC)
the final link with -fPIE will fail. It's convenient to
be able to do this for building against an uninstalled libopus.

Unfortunately this involves moving the -fPIE check until
after the library checks, so we can add it to the link
line before testing. We also have to reference a libopus
symbol to keep the linker from skipping the dependency.