Fix for Opusdec output at 44.1kHz adding a pop at ~1 second.
Gainless, 2012, and Ivqcl on Hydrogenaudio reported and helped
identify an issue where opusdec was adding a click in 44.1kHz
output. This was becaue the loop needed to feed the resampler
for >1024 sample input wasn't accounting for the preskip and
was putting too many samples into the resampler on the first
packet. Eventually the output got ahead of the container
granpos and the decoder dropped some samples to catch up.

For 1 second pages, the opusenc default, dropped samples
happened just about at the 1 second mark.