When simulating packet loss in opusdec, use the real durations.
In Ogg when you 'lose' packets within a page (something that isn't actually
possible) you can also become unable to determine their durations.
Previously the code would just let the durations be wrong, which would
potentially trigger whining about the granpos not matching the number
of decoded samples. This patch makes it pull the durations from the
original frames. This sould actually make the simulation closer to
the RTP behavior where the packets are all timestamped in any case.