Remove --speech and --music arguments from opusenc.
All these options do is bias the automatic mode selection in libopus,
but it seems that many people believe they select between LP and
MDCT modes, and it's very hard to convince them otherwise. As is they
aren't very useful, and they're creating a lot of confusion which
is distracting people from actually useful testing.