ifdef-protect the pcap code.
We don't want to break the build when libpcap isn't available,
and eventually we want this tool to send/receive rtp streams
as well as sniff them, so move the pcap-related code to a
separate function, which we only call if it's available.

This means that by default the programme does nothing.