Reject playback of multichannel files through sys/soundcard.h.
Previously we wrote out the n channel data to an fd configured
for mono output, which doesn't sound very good. Better to reject
it.

There is a SNDCTL_DSP_CHANNELS we can use to request multichannel
output, but there doesn't seem to be a way to set the channel
mapping, so it's not much better.