Document the intent of the #ifdefs around query_cpu_support.
I found the code bypassing the query definition both when
SSE isn't available _and_ on x86_64, where it is unconditionally
available, confusing. That's not what I expected from general
cpuid reporting implementation.

Instead, it's specialized to the code's use only to warn before
the user hits an invalid instruction, so it makes sense to disable
it in those particular cases.

Thanks to Greg Maxwell for explaining what was going on.