Add the '=' separator inside comment_add().
This avoids having to remember to pass it in as part of the
tag name, which is confusing when the fuction takes both
a tag and a value as separate arguments.

The separator is only added if the tag is non-null, so
the --comment option, which passes tag and value together
in the value argument still works fine.