More autoconf housekeeping
Fix the libm exp() test to not unconditionally add it to everything via
LIBS, and fix opus{enc,dec}_LDADD to actually use LIBM that it sets
rather than hardcoding it there as well.

Use AS_HELP_STRING to format all option descriptions.

Get rid of a bunch of shadow variables that aren't really needed.

Drop the x warts in tests, it really isn't needed, and there are lots
of places we don't use it, so if it was they'd already be exploding.

Let the pkg-config tests actually fail if they don't find the desired
package, rather than also running the manual test which will just fail
again in that case.

Don't overlink flac with -lm in the manual test.  That was fixed in
its .pc in commit 5da86630bee2336e251f0a140f33332af703c43e.

Change --disable-flac to --without-flac.  The normal autoconf idiom is
to use 'enable' for intrinsic features and 'with' for external software
support that is optional.

Change Opus_{CFLAGS,LIBS} to OPUS_{CFLAGS,LIBS}, since that what opus.m4
will try to set if pkg-config isn't available.

Drop the check for __malloc_hook, we don't use it anywhere here.