Get rid of LIBM altogether
And the exp() check too.  Nothing in the code actually tests for that
and it's irrelevant whether it needs libm or not, since plenty of other
things definitely do need it.

So add the hardcoded -lm back where it is definitely needed (at the
very least for the speex resampler), and make sure it is in LIBS when
we test for lrint and f{min,max}f, since those tests do fail on Linux
if libm isn't linked.