Tweak the sound driver detection tests
Test for the available possibilities in the code in the same order that
we use when including the headers.  This ensures that BSDs which test
positive for both sndio and oss compat won't include the headers for one
and the code for the other.  The sndio support was added in commit
064bb2176518117531e0032747128761a19e5439, but only preferred sndio when
including the headers.

Enable the OSS code for all the possible HAVE_*_SOUNDCARD_H options.
A more comprehensive test and additional macros were added in commit
a0c500527daa6a8638eabac382810b1c42f442a2, but only to include the
headers, not to also enable the relevant code.