opusinfo: Add detection of extra packets in header pages.
Additional header packets are disallowed by the spec and
rejected by the opusfile decoder library. Report this
condition in opusinfo so it's more obvious what's wrong
with such files.

The output is a little confusing when the extra packet
is in the first page as it also tries to parse it as
a comment header.

Signed-off-by: Ralph Giles <giles@thaumas.net>