opusrtp: Clean up tag header style.
Regularize whitespace and be consistent using a fixed 8-byte
length for the `OpusTags` packet header, rather can calling
strlen on the literal. This length, and the offset of the
subsequence fields, is fixed by the specification, so using
a fixed length would great a less-corrupt file if the header
sequence were somehow changed.

Thanks to willson-chen for pointint out the style issue,
and Mark Harris for recommending the fixed header length.